Renamed 'database' to 'ytdb'
Renamed the 'database' folder to 'ytdb' to minimize confusion with instantiation with the same name. Also changed the name of 'youtube_reader.py' to 'reader.py' since it is now clearly associated with the ytdb module.
