+++ ytdb/load.py
@@ -3,7 +3,7 @@
 
 from tinydb import TinyDB
 
-from database.storage import OnlineJSONStorage
+from ytdb.storage import OnlineJSONStorage
 
 YT_DATABASE_URL = "https://squeemos.pythonanywhere.com/static/youtube.json"
 
@@ -18,13 +18,13 @@ def _docstring_parameter(*sub):
 
 @_docstring_parameter(YT_DATABASE_URL)
 def from_url(data_url: str = YT_DATABASE_URL):
-    """Convenience function for loading a TinyDB database from a web address.
+    """Convenience function for loading a TinyDB ytdb from a web address.
 
     As of now, only supports loading databases in JSON format.
 
     Parameters
         data_url: Online address at which TinyDB data is retrievable.
-            Defaults to YT_DATABASE_URL defined at top of 'database.load'.
+            Defaults to YT_DATABASE_URL defined at top of 'ytdb.load'.
             YT_DATABASE_URL = {0}
     """
     return TinyDB(data_url, storage=OnlineJSONStorage)
